Move StatusIndicator out of Latex clas
