+++ lib/composer.js
@@ -60,7 +60,7 @@ export default class Composer {
       })
     }
 
-    latex.setStatus('LaTeX Build', 'highlight', 'sync', true, 'Click to kill LaTeX build.', () => latex.process.killChildProcesses())
+    latex.status.show('LaTeX Build', 'highlight', 'sync', true, 'Click to kill LaTeX build.', () => latex.process.killChildProcesses())
 
     latex.log.group('LaTeX Build')
     const rootFilePath = this.resolveRootFilePath(filePath)
