+++ lib/latex.js
@@ -4,6 +4,7 @@ import fs from 'fs-plus'
 import _ from 'lodash'
 import { heredoc } from './werkzeug'
 import ProcessManager from './process-manager'
+import StatusIndicator from './status-indicator'
 
 function defineDefaultProperty (target, property) {
   const shadowProperty = `__${property}`
@@ -23,6 +24,7 @@ function defineDefaultProperty (target, property) {
 
 export default class Latex {
   process = new ProcessManager()
+  status = new StatusIndicator()
 
   constructor () {
     this.createLogProxy()
@@ -38,32 +40,6 @@ export default class Latex {
 
   setLogger (logger) {
     this.logger = logger
-  }
-
-  attachStatusBar (statusBar) {
-    const StatusIndicator = require('./views/status-indicator')
-    this.statusIndicator = new StatusIndicator({ type: 'idle' })
-    this.statusTile = statusBar.addRightTile({
-      item: this.statusIndicator,
-      priority: 9001
-    })
-  }
-
-  detachStatusBar () {
-    if (this.statusTile) {
-      this.statusTile.destroy()
-      this.statusTile = null
-    }
-    if (this.statusIndicator) {
-      this.statusIndicator.destroy()
-      this.statusIndicator = null
-    }
-  }
-
-  setStatus (text, type, icon, spin, title, onClick) {
-    if (this.statusIndicator) {
-      this.statusIndicator.update({ text, type, icon, spin, title, onClick })
-    }
   }
 
   getDefaultLogger () {
