+++ lib/loggers/default-logger.js
@@ -37,7 +37,7 @@ export default class DefaultLogger extends Logger {
 
   updateStatus () {
     const icon = this.view.isVisible() ? 'chevron-down' : 'chevron-up'
-    latex.setStatus('LaTeX Log', this.type, icon, false, 'Click to toggle LaTeX log.', () => this.toggle())
+    latex.status.show('LaTeX Log', this.type, icon, false, 'Click to toggle LaTeX log.', () => this.toggle())
   }
 
   showErrorMarkers (messages) {
