+++ lib/main.js
@@ -47,9 +47,9 @@ export default {
 
   consumeStatusBar (statusBar) {
     this.bootstrap()
-    latex.attachStatusBar(statusBar)
+    latex.status.attachStatusBar(statusBar)
     return new Disposable(() => {
-      latex.detachStatusBar()
+      latex.status.detachStatusBar()
     })
   },
