+++ lib/views/status-label.js
@@ -3,7 +3,7 @@
 
 import etch from 'etch'
 
-export default class StatusIndicator {
+export default class StatusLabel {
   constructor (properties) {
     this.properties = properties
     etch.initialize(this)
